create common component for dark screens
