+++ ts/components/screens/DarkLayout.tsx
@@ -3,6 +3,7 @@ import * as React from "react";
 import { ImageSourcePropType, StyleProp, ViewStyle } from "react-native";
 import { StyleSheet } from "react-native";
 import customVariables from "../../theme/variables";
+import AnimatedScreenContent from "./AnimatedScreenContent";
 import ScreenContent from "./ScreenContent";
 import TopScreenComponent from "./TopScreenComponent";
 
@@ -17,6 +18,10 @@ type Props = Readonly<{
   bounces?: boolean;
   topContent?: React.ReactNode;
   hasDynamicSubHeader?: boolean;
+  dynamicSubHeader?: React.ReactNode;
+  topContentHeight?: number;
+  footerContent?: React.ReactNode;
+  contextualHelp?: { title: string; body: () => React.ReactNode };
 }>;
 
 const styles = StyleSheet.create({
@@ -27,8 +32,19 @@ const styles = StyleSheet.create({
 });
 
 export default class DarkLayout extends React.Component<Props> {
+  private screenContent() {
+    return (
+      <React.Fragment>
+        <View style={styles.headerContents}>
+          <View spacer={true} />
+          {this.props.topContent}
+        </View>
+
+        {this.props.children}
+      </React.Fragment>
+    );
+  }
   public render() {
-    const { title } = this.props;
     return (
       <TopScreenComponent
         goBack={this.props.allowGoBack}
@@ -36,21 +52,38 @@ export default class DarkLayout extends React.Component<Props> {
         dark={true}
         headerBody={this.props.headerBody}
         appLogo={this.props.appLogo}
+        contextualHelp={this.props.contextualHelp}
       >
-        <ScreenContent
-          hideHeader={this.props.hideHeader}
-          title={title}
-          icon={this.props.icon}
-          dark={true}
-          contentStyle={this.props.contentStyle}
-          bounces={this.props.bounces}
-        >
-          <View style={styles.headerContents}>
-            <View spacer={true} />
-            {this.props.topContent}
-          </View>
-          {this.props.children}
-        </ScreenContent>
+        {this.props.hasDynamicSubHeader ? (
+          <AnimatedScreenContent
+            hideHeader={this.props.hideHeader}
+            title={this.props.title}
+            icon={this.props.icon}
+            dark={true}
+            contentStyle={this.props.contentStyle}
+            dynamicSubHeader={this.props.dynamicSubHeader}
+            topContentHeight={
+              this.props.topContentHeight ? this.props.topContentHeight : 0
+            }
+            animationOffset={40}
+          >
+            {this.screenContent()}
+          </AnimatedScreenContent>
+        ) : (
+          <ScreenContent
+            hideHeader={this.props.hideHeader}
+            title={this.props.title}
+            icon={this.props.icon}
+            dark={true}
+            contentStyle={this.props.contentStyle}
+            bounces={this.props.bounces}
+          >
+            {this.screenContent()}
+          </ScreenContent>
+        )}
+        {this.props.footerContent && (
+          <View footer={true}>{this.props.footerContent}</View>
+        )}
       </TopScreenComponent>
     );
   }
